Adding tasklet for bottom half
In the top half of interrupt process, the system interrupt must be
disabled
Since generic_handle_irq invoke interrupt manually and the interrupt
handler will run in the context of kernel process not in the hardware
interrupt context. In the hardware interrupt context, interrupt will be
disable and stack will be switched automatically.